vectorMatch/requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
filelock==3.16.0
fsspec==2024.9.0
huggingface-hub==0.24.7
idna==3.8
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
packaging==24.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
sympy==1.13.2
tokenizers==0.19.1
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
urllib3==2.2.3
h11~=0.14.0
pip~=22.3.1
sniffio~=1.3.1
httpcore~=1.0.7
setuptools~=65.5.0
pydantic~=2.10.2
ollama~=0.4.2
httpx~=0.27.2
anyio~=4.6.2.post1
pydantic_core~=2.27.1